* Transitioned from Boost C++'s ptree to the PugiXML library, for storing options and download information. * Cleaned up some code in a few places.

+++ src/cmnroutines.hpp
@@ -46,7 +46,6 @@
 #include "default_var.hpp"
 #include <boost/optional.hpp>
 #include <QDateTime>
-#include <exception>
 #include <string>
 #include <cstdio>
 #include <QString>
@@ -92,7 +91,7 @@ public:
     struct CurlInfoExt {
         bool status_ok;            // Whether 'CURLE_OK' was returned or not
         std::string status_msg;    // The status message, if any, returned by the libcurl functions
-        long response_code;        // The HTTP/FTP response code
+        long long response_code;   // The HTTP/FTP response code
         double elapsed;            // Total time in seconds for the previous transfer (including name resolving, TCP connect, etc.)
         std::string effective_url; // In cases when you've asked libcurl to follow redirects, it may very well not be the same value you set with 'CURLOPT_URL'
         double content_length;     // The size of the download, i.e. content length
@@ -112,8 +111,8 @@ public:
     std::string findCfgFile(const std::string &cfgFileName);
     std::vector<CurlDlInfo> readDownloadInfo(const std::string &xmlCfgFile = CFG_HISTORY_FILE);
     bool writeDownloadInfo(CurlDlInfo dl_info_list, const std::string &xmlCfgFile = CFG_HISTORY_FILE);
-    short convDlStat_toInt(const GekkoFyre::DownloadStatus &status);
-    GekkoFyre::DownloadStatus convDlStat_toEnum(const short &s);
+    int convDlStat_toInt(const GekkoFyre::DownloadStatus &status);
+    GekkoFyre::DownloadStatus convDlStat_toEnum(const int &s);
     QString convDlStat_toString(const GekkoFyre::DownloadStatus &status);
 
     CurlInfo verifyFileExists(const QString &url);
